+++ src/test/scala/nl/knaw/dans/easy/pid/generator/DOIGeneratorSpec.scala
@@ -22,7 +22,7 @@ import scala.util.{ Failure, Success }
 
 class DOIGeneratorSpec extends SeedDatabaseFixture
   with PropertiesSupportFixture
-  with DOIGeneratorComponent
+  with DOIGeneratorWiring
   with PropertiesComponent
   with SeedStorageComponent
   with DatabaseComponent
